+++ src/config/tabs.rs
@@ -55,7 +55,8 @@ pub enum PaneTypeFile {
     },
 }
 
-#[derive(Debug, Clone, Hash, Eq, PartialEq, strum::Display)]
+#[derive(Debug, Clone, Copy, Hash, Eq, PartialEq, strum::Display, strum::EnumDiscriminants)]
+#[strum_discriminants(derive(strum::Display, Hash))]
 pub enum PaneType {
     Queue,
     #[cfg(debug_assertions)]
@@ -80,30 +81,33 @@ pub enum PaneType {
     },
 }
 
+pub const PANES_ALLOWED_IN_BOTH_TAB_AND_LAYOUT: [PaneTypeDiscriminants; 1] =
+    [PaneTypeDiscriminants::Property];
+
 #[cfg(debug_assertions)]
-pub const UNFOSUSABLE_TABS: [PaneType; 7] = [
-    PaneType::AlbumArt,
-    PaneType::Lyrics,
-    PaneType::ProgressBar,
-    PaneType::Header,
-    PaneType::Tabs,
-    PaneType::TabContent,
-    PaneType::FrameCount,
+pub const UNFOSUSABLE_TABS: [PaneTypeDiscriminants; 7] = [
+    PaneTypeDiscriminants::AlbumArt,
+    PaneTypeDiscriminants::Lyrics,
+    PaneTypeDiscriminants::ProgressBar,
+    PaneTypeDiscriminants::Header,
+    PaneTypeDiscriminants::Tabs,
+    PaneTypeDiscriminants::TabContent,
+    PaneTypeDiscriminants::FrameCount,
 ];
 
 #[cfg(not(debug_assertions))]
-pub const UNFOSUSABLE_TABS: [PaneType; 6] = [
-    PaneType::AlbumArt,
-    PaneType::Lyrics,
-    PaneType::ProgressBar,
-    PaneType::Header,
-    PaneType::Tabs,
-    PaneType::TabContent,
+pub const UNFOSUSABLE_TABS: [PaneTypeDiscriminants; 6] = [
+    PaneTypeDiscriminants::AlbumArt,
+    PaneTypeDiscriminants::Lyrics,
+    PaneTypeDiscriminants::ProgressBar,
+    PaneTypeDiscriminants::Header,
+    PaneTypeDiscriminants::Tabs,
+    PaneTypeDiscriminants::TabContent,
 ];
 
 impl Pane {
     pub fn is_focusable(&self) -> bool {
-        !UNFOSUSABLE_TABS.contains(&self.pane)
+        !UNFOSUSABLE_TABS.contains(&self.pane.into())
     }
 }
 
@@ -456,11 +460,14 @@ pub(crate) fn validate_tabs(layout: &SizedPaneOrSplit, tabs: &Tabs) -> Result<()
         .flat_map(|tab_pane| {
             layout_panes.iter().filter(|layout_pane| layout_pane.pane == tab_pane.pane)
         })
+        .filter(|pane| !PANES_ALLOWED_IN_BOTH_TAB_AND_LAYOUT.contains(&pane.pane.into()))
+        .map(|pane| PaneTypeDiscriminants::from(pane.pane))
+        .unique()
         .collect_vec();
     ensure!(
         panes_in_both_tabs_and_layout.is_empty(),
         "Panes cannot be in layout and tabs at the same time. Please remove following tabs from either layout or tabs: {}",
-        panes_in_both_tabs_and_layout.iter().map(|pane| &pane.pane).dedup().join(", ")
+        panes_in_both_tabs_and_layout.iter().join(", ")
     );
 
     Ok(())