+++ docs/astro.config.mjs
@@ -63,7 +63,7 @@ export default defineConfig({
                         ],
                     },
                     {
-                        label: "Release (v0.7.0)",
+                        label: "Release (v0.8.0)",
                         link: "release/overview",
                         badge: { text: "Stable", variant: "default" },
                         icon: "seti:todo",
@@ -102,6 +102,50 @@ export default defineConfig({
                                 label: "Theme gallery",
                                 autogenerate: {
                                     directory: "release/themes",
+                                },
+                            },
+                        ],
+                    },
+                    {
+                        label: "Release (v0.7.0)",
+                        link: "release-0-7-0/overview",
+                        badge: { text: "Stable", variant: "default" },
+                        icon: "seti:todo",
+                        items: [
+                            {
+                                label: "Overview",
+                                link: "release-0-7-0/overview",
+                            },
+                            {
+                                label: "Installation",
+                                link: "release-0-7-0/installation",
+                            },
+                            {
+                                label: "Try without installing",
+                                link: "release-0-7-0/try-without-install",
+                            },
+                            {
+                                label: "Configuration",
+                                autogenerate: {
+                                    directory: "release-0-7-0/configuration",
+                                },
+                            },
+                            {
+                                label: "Guides",
+                                autogenerate: {
+                                    directory: "release-0-7-0/guides",
+                                },
+                            },
+                            {
+                                label: "Reference",
+                                autogenerate: {
+                                    directory: "release-0-7-0/reference",
+                                },
+                            },
+                            {
+                                label: "Theme gallery",
+                                autogenerate: {
+                                    directory: "release-0-7-0/themes",
                                 },
                             },
                         ],