+++ docs/src/content/docs/release-0-7-0/data.ts
@@ -1,0 +1,2 @@
+export const BASE_PATH = "/rmpc/release-0-7-0";
+export const path = (path: string) => `${BASE_PATH}/${path}`;