+++ CHANGELOG.md
@@ -32,6 +32,7 @@ They have additional config properties compared to their predecesors.
 - Status messages will now disappar automatically even when idle
 - Lyrics should now sync better because they are now scheduled precisely instead of periodically
 - MSRV to 1.82
+- Song metadata is now split into groups
 
 ### Fixed
 
@@ -46,6 +47,7 @@ They have additional config properties compared to their predecesors.
 - AddToPlaylist not working for local songs
 - rmpc waiting potentionally forever for MPD's response
 - Adding songs which do not belong to any album not working in `Artists` and `AlbumArtists` panes not working
+- Songs metadata not being sorted in preview column
 
 ### Deprecated