+++ CHANGELOG.md
@@ -50,6 +50,7 @@ They have additional config properties compared to their predecesors.
 - Adding songs which do not belong to any album not working in `Artists` and `AlbumArtists` panes not working
 - Songs metadata not being sorted in preview column
 - Prevent album art rendering when modal is open
+- Fix panic when `ProgressBar` pane had insufficient height
 
 ### Deprecated