+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+
 ## [Unreleased]
+
+### Added
+
+### Changed
+
+### Fixed
+
+## [0.8.0] - 2025-02-16
 
 ### Added
 
@@ -254,7 +263,8 @@ They have additional config properties compared to their predecesors.
 
 ## [0.1.0] - 2024-06-21
 
-[unreleased]: https://github.com/mierak/rmpc/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/mierak/rmpc/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/mierak/rmpc/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/mierak/rmpc/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/mierak/rmpc/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/mierak/rmpc/compare/v0.4.0...v0.5.0